Final Version
No I won't be adding the previous ones. There are a total of 16 iterations half of which are testing the reason for the original button failure. In the end, rebuilding it from scratch seems to have solved the problem, even without any debugging on my end. So we're probably blaming MOJO IDE and LUCID HDL.